requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
python-multipart
pandas
python-dateutil
pydantic
httpx